Support conversion of Python instances deriving from native objects
If a Python type derives from a ClassWrapper type allow an instance of
the Python type to be converted into a ClassWrapper instance.
This allows for example to pass the Python instance to a native function
which has an argument of the native type from which the Python one derives.